backend/pyproject.toml
[project]
name = "bourbaki-backend"
version = "0.1.0"
description = "An autonomous agent for mathematical reasoning and proof — Python backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-ai>=1.55.0",
    "pydantic-settings>=2.0",
    "sympy>=1.13",
    "httpx>=0.27",
    "python-frontmatter>=1.1.0",
    "sse-starlette>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
